+++ exe/merger
@@ -1,3 +1,3 @@
 #!/usr/bin/env ruby
-require 'merger/cli'
-Merger::CLI.start
+require 'filemerger/cli'
+Filemerger::CLI.start
