+++ lib/filemerger/version.rb
@@ -1,0 +1,3 @@
+module Filemerger
+  VERSION = "0.3.0"
+end
