--- lib/merger/version.rb
@@ -1,3 +1,0 @@
-module Merger
-  VERSION = "0.2.1"
-end
